backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
google-generativeai
markdown
beautifulsoup4